chore: satisfy CI clippy (Rust 1.91) across core, api, desktop
Test modules moved to the end of their files, cheap or_else chains flattened,
is_multiple_of / is_none_or, unnecessary casts in tests, a duplicated cfg
attribute on the Linux pulse router.

+++ crates/paracord-api/tests/message_delete_receipts.rs
@@ -173,8 +173,8 @@ async fn overwrite(f: &Fixture, allow: Permissions, deny: Permissions) {
         f.channel,
         f.member_id,
         1,
-        allow.bits() as i64,
-        deny.bits() as i64,
+        allow.bits(),
+        deny.bits(),
     )
     .await
     .unwrap();
